small update, starting on switch. might have to change entire output method
switch not only has uppercase but also a different keyboard entirely for punctuation. I've been able to make the way i've been doing work because of Character.Uppercase etc but maybe i'm going to have to dumb down the way I do it so that each button press isn't directly shown but instead just the whole path is. would be a lot simpler but not as cool.